requirements.txt
flask
Flask-Cors
requests
pillow